+++ futurefinity/__init__.py
@@ -15,6 +15,6 @@
 #   See the License for the specific language governing permissions and
 #   limitations under the License.
 
-__version__ = ("0", "1", "1", "1")
+__version__ = ("0", "2", "0", "-1000")
 
-version = "0.1.1"
+version = "0.2.0.dev0"