+++ futurefinity/interface/template.py
@@ -15,8 +15,6 @@
 #   See the License for the specific language governing permissions and
 #   limitations under the License.
 
-from futurefinity.utils import *
-
 try:
     import jinja2
 except ImportError: