+++ futurefinity/server.py
@@ -22,7 +22,8 @@ right RequestHandler and make response to client.
 """
 
 from futurefinity.utils import ensure_str, ensure_bytes
-from futurefinity.protocol import HTTPHeaders, HTTPRequest, HTTPError
+from futurefinity.protocol import (HTTPHeaders, HTTPRequest,
+                                   HTTPResponse, HTTPError)
 
 import futurefinity
 
@@ -197,52 +198,36 @@ class HTTPServer(asyncio.Protocol):
             app=self.app,
             server=self,
             request=request,
-            make_response=self.make_response
+            respond_request=self.respond_request
         )
         await request_handler.handle(**matched_obj)
 
-    def make_response(self, status_code: int,
-                      response_headers: HTTPHeaders,
-                      response_body: bytes):
+    def respond_request(self, request: HTTPRequest, response: HTTPResponse):
         """
         Make http response to client.
         """
         if self.http_version == 20:
             pass  # HTTP/2 will be implemented later.
         else:
-            self.make_http_v1_response(status_code, response_headers,
-                                       response_body)
+            self.respond_http_v1_request(request, response)
 
-    def make_http_v1_response(
-        self,
-        status_code: int,
-        response_headers: HTTPHeaders,
-        response_body: bytes):
-
+    def respond_http_v1_request(self, request: HTTPRequest,
+                                response: HTTPResponse):
         """
         Make HTTP/1.x response to client.
 
         This function should not be called directly, make_response() function
         will handle it to right http version.
         """
-        response_text = b""
-        if self.http_version == 10:
-            response_text += b"HTTP/1.0 "
-        elif self.http_version == 11:
-            response_text += b"HTTP/1.1 "
+        use_keep_alive = (self.http_version == 11 and
+                          self.app.settings.get("allow_keep_alive", True))
 
-        response_text += ensure_bytes(str(status_code)) + b" "
+        if use_keep_alive and "keep-alive" not in response.headers:
+            response.headers.add("keep-alive", "timeout=100, max=100")
 
-        response_text += ensure_bytes(http.client.responses[
-            status_code]) + b"\r\n"
-        for (key, value) in response_headers.get_all():
-            response_text += ensure_bytes("%(key)s: %(value)s\r\n" % {
-                "key": key, "value": value})
-        response_text += b"\r\n"
-        response_text += ensure_bytes(response_body)
-        self.transport.write(response_text)
-        if self.http_version == 11 and self.app.settings.get(
-         "allow_keep_alive", True):
+        self.transport.write(response.make_http_v1_response())
+
+        if use_keep_alive:
             self.reset_server()
         else:
             self.transport.close()